# multas-backend/docs/API_OPENAPI.yaml
openapi: 3.1.0
info:
  title: PMT Multas API - TANDAS 0–8, 10 y 11
  version: 1.2.0
  description: Contrato de autenticación, operación municipal, finanzas, solvencias, analítica, notificaciones y migración histórica controlada.
servers:
  - url: /api/v1
...
paths:
...
  /institutional-rules:
    get:
      tags: [Institutional rules]
      security: [{ sessionCookie: [] }]
      summary: Lista versiones de reglas legales y económicas
      responses: { '200': { description: Reglas versionadas } }
    post:
      tags: [Institutional rules]
      security: [{ sessionCookie: [] }]
      summary: Crea una nueva vigencia y cierra la anterior sin sobrescribirla
      responses: { '201': { description: Versión creada }, '409': { $ref: '#/components/responses/Error' } }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pmt_session
  responses:
    Error:
      description: Error estable con requestId
      content:
        application/json:
          schema:
            type: object
            required: [error, meta]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
              meta:
                type: object
                required: [requestId]
                properties:
                  requestId: { type: string, format: uuid }